feat: ApplyAttributeEffect

+++ zombies/src/main/java/org/phantazm/zombies/event/entity/EntityAttributeModifierRemoveEvent.java
@@ -35,8 +35,8 @@ public class EntityAttributeModifierRemoveEvent implements EntityInstanceEvent, 
     }
 
     @Override
-    public float attributeAmount() {
-        return (float) removedModifier.getAmount();
+    public double attributeAmount() {
+        return removedModifier.getAmount();
     }
 
     @Override
